src/requirements.txt
psycopg2-binary
python-dotenv
Flask
flask-cors
